Fix time units (original assumption due to poor fact checking was that the time unit was ms)

+++ idserver.h
@@ -73,7 +73,7 @@ private:
 	 */
     int ticStart;
     int ticLoop;
-    int ticTurn;
+    double ticTurn;
     MapType ticNodes;
 };
